REST/requirements.txt
fastapi==0.99.1
uvicorn[standard]==0.22.0
alembic==1.11.1
aioredis==2.0.1
asyncpg==0.27.0
psycopg2-binary==2.9.6
tortoise-orm[asyncpg]
python-multipart
